Switch errorcode for too big namespace declaration and fix underflow test
Summary:
I describe the issue a bit in D63361167 as well, but basically the problem is that because we read all the objects from one continuous buffer, there is a case where underflow in one object will still leave enough data to begin parsing a new object. If, for example, the last field in one object is a string and it underflows, the cursor will not be advanced at all when the parser detects the underflow. The test will advance the cursor one byte, but this leaves the cursor at some arbitrary byte within the buffer.

In the past this was fine as it just immediately underflowed or read some random string into the field for the next object, but with the namespace tuple it more often than not will read a tuple item count that is > 32 items, causing a non-underflow error and failing the test.

Reviewed By: afrind

Differential Revision: D63636643

fbshipit-source-id: 1847571df9c1c740af77a4bfe9c8590c23f358ff