Implement track namespace tuple
Summary:
Apologies for the large diff. It turns out namespace is used in a lot of places, and changing it changes a lot of stuff.

Track namespace has been redefined as an N-tuple:

> Track namespace is an ordered N-tuple of bytes where N can be between 1 and 32.  The structured nature of Track Namespace allows relays and applications to manipulate prefixes of a namespace. Track name is a sequence of bytes.

Where a tuple is defined by:
> x (tuple):  Indicates that x is a tuple, consisting of a variable length integer encoded as described in ([RFC9000], Section 16), followed by that many variable length tuple fields, each of which are encoded as (b) above.

I put the namespace in its own struct as there's a lot of places where we log the value of namespace, and the vector to string conversion is mildly annoying.

Reviewed By: afrind

Differential Revision: D63361167

fbshipit-source-id: 02e990b291213203a7324bd7018591cce8f54063