Implement draft-06 stream types and subgroups
Summary:
Splitting DATAGRAM and STREAM_HEADER out of FrameType required a refactor of MoQCodec that was a long time coming.

Now there's a pure virtal base class, a MoQControlCodec and MoQObjectStreamCodec (works for datagrams).

Removing ForwardPreference::Object was painful, because stream-per-object now requires an 'end-of-group' Status to terminate it (at least for now, see comments here: https://github.com/moq-wg/moq-transport/issues/442

For original publishers, there is now a 'publishStreamPerObject' API that will do it all for you.  It does violate the spec by re-using the same object ID for the status.  Relays will end up proxying as normal SubGroup, and hopefully the OP included END_OF* to terminate the stream.

Reviewed By: yuandagits

Differential Revision: D63567571

fbshipit-source-id: 9efbe8a4497fdc28d9eaef81fa3c0611e9581b4c