Add length to control messages
Summary:
This adds an explicit length to all control messages.  It's encoded as a varint, which is annoying.  moxygen always writes it as a 2 byte varint, and fails writing control messages longer than 2^14.

This removes all the calls to cursor.totalLength() from the control stream, which was a bit of a DoS vector.  There's still one on the object-stream path, which we can work on removing as we iterate.

This diff started as a backout of D57702981.

Reviewed By: yuandagits

Differential Revision: D63587022

fbshipit-source-id: df291d4d289efc69a645547605e81ec5860f6c43