MoQ Relay implement SUBSCRIBE_NAMESPACE
Summary:
Replace the `map<namespace, session>` with a tree organized by namespace tuple field.  Each node has the following fields:

`name` - the name of the tuple field for this node
 `sourceSession` - the session that has announced the namespace matching the path from `announceRoot_` to that node
`children` - vector of nodes that have this node as a prefix
`sessions` - vector of sessions that have issued a `SUBSCRIBE_NAMESPACE` for the prefix matching the path to this node

With that structure, the following operations are defined:

* `onAnnounce` - create the nodes along the path to the announced namespace and set `sourceSession`.  Collect sessions from all prefix nodes and forward the Announce to them
* `onUnannounce` - find the node for the announced session and clear `sourceSession`.  Collect sessions from all prefix nodes and forward the Unannounce to them
* `onSubscribeNamespace` - create the nodes along the path to the prefix and add session to the list of sessions.  Enumerate all announcements below this node in the tree and forward them to the session.
* `onUnsubscribeNamespace` - find the node for the prefix and remove session from the list of sessions.

When a session goes away, remove it from all locations in the tree where it might have existed.

There are a number of TODOs.  Having vectors in the struct are not particularly efficient - should probably be maps.

Support for > 1 ANNOUNCE per namespace is not yet supported

Reviewed By: yuandagits

Differential Revision: D63717147

fbshipit-source-id: 07262c6e0f95ef59a588ae4ff8a1e2393cd2c70b