Fix memory error in controlMessages, test cleanup for OSS
Summary:
UnboundedQueue will attempt to dequeue an item even if it exits from sem_.wait() with a cancellation exception.  This can cause a UAF which I previously attempted to fix in D59649785.  Instead, hold a shared_ptr in controlMessages.

The minor test cleanup is for OSS -- sometimes it would complain about SingleonVault initialization for coro::timeout.

Reviewed By: NEUDitao

Differential Revision: D67601733

fbshipit-source-id: 91533d398f3858141e0fcf0ebc7e95cfda648e19