Remove ForwardPreference from ObjectHeader
Summary: This is a vestige of the old API and draft versions.  draft-07 still defines datagram vs subgroup preferences as an immutable, publisher selected property.  However, there isn't necessarily need to cache this, because cached objects are retrieved via FETCH, on a single stream, which discards the preference.  Forwarding the preferrnce in moxygen is a matter of using the correct API (datagram() vs objectStream() or beginSubgroup()).

Reviewed By: NEUDitao

Differential Revision: D67906300

fbshipit-source-id: 06a37f8bbce43b78971059b488e30d4f50db7bea