Fixed bug reading size of IOBuf
Summary: - Fixed getting wrong IOBuf length. In a lot o places I was assuming IOBuf was a plain buffer, and this is TOTALLY WRONG. To get the length of a NON coalesced IOBuf we should use `computeChainDataLength`, then use `cursor` to get the all the data. Finally use `coalesced` ONLY when is totally necessary to all data in a continuous mem space

Reviewed By: afrind

Differential Revision: D67837059

fbshipit-source-id: 22cb85a0bb52fd7748e7c167b3089ddf1fc72389