Publisher and Subscriber
Summary:
This is the beginning of rewriting the control interfaces of MoQSession.  The goal is to have the interface match the key verbs of the protocol.  Currently:

Publisher:

  trackStatus
  subscribe
  fetch
  subscribeAnnounces

Subscriber:

  announce

Both:

  `goaway

With the exception of goaway, which has no response, each of these APIs is a coroutine.  The return value is a "Handle" to the corresponding state that supports the operations allowed (eg: SubscribeHandle = subscribeUpdate, unsubscribe).

The API is designed to be symmetric, such that MoQSession will implement these APIs for the caller, and will be given a callback to handle these APIs when they arrive via the network.

Reviewed By: sharmafb

Differential Revision: D67865127

fbshipit-source-id: 3cd144664eec78f1cbbf0683b3f6fdf861f943c6