Create a WebTransportImpl::DeliveryCallback
Summary:
`proxygen/lib/http/webtransport/WebTransport.h` is the only file that contains the relevant change. In that file, I'm:
* Creating an abstract `DeliveryCallback` class.
* Modifying `WebTransport::writeStreamData` and `WebTransport::StreamWriteHandle::writeStreamData` to also take in a `DeliveryCallback*` parameter.

In all other files, I'm just modifying callsites to pass in a `nullptr`. There's no functional change in this diff (i.e. we're not doing anything with the passed-in `DeliveryCallback*`). Future diffs contain functional changes.

Reviewed By: hanidamlaj

Differential Revision: D69617977

fbshipit-source-id: 1e9c2b4086a5b2cf0537eda844d0cc6f59e44282