Adding streamCount to SUBSCRIBE_DONE (#25)
Summary:

This allows the session to delay delivering SUBSCRIBE_DONE at least until all streams related to that subscription have been opened at the receiver.  This also needs a timeout for malicious or buggy peers, or those that reset a stream without a reliable offset > stream header.  This will be added later.

The API is somewhat awkward since the caller has to fill in that part of the struct, but it's overwritten by the session.  One possibility is to make two SubscribeDone structs

```
SubscribeDone { // API version
  Fields
};

SubscribeDoneDetail : public SubscribeDone { // Wire version
  uint64_t streamCount;
}
```

Reviewed By: sharmafb

Differential Revision: D69501084