Fix fetch input validation to allow single group requests
Summary:
Fetch parameters are defined as:
> EndObject: The end Object ID, plus 1. A value of 0 means the entire group is requested.

This means that to request a part of a single group you would request (for example) 1.5 -> 1.0. This would fail the validation, as the end location is less than the start location, but is actually a valid request.

Also updated the fetch error test case as that case actually shouldn't be an error.

Reviewed By: sharmafb

Differential Revision: D70138599

fbshipit-source-id: 83df50db7ea133fb655d78ecd7253ad77ae4cc1b