Add version fields to the MoQFrameParser and MoQFrameWriter
Summary: This commit has no functional change. We could potentially have the initialization done in the constructor instead, but this can may things a little more messy for objects like the MoQControlCodec which must be initialized before the version is known (if we go that route, we'd likely need to have a pointer to a parser instead, and initialize it after the version is known).

Reviewed By: afrind

Differential Revision: D70864844

fbshipit-source-id: 12d509e50385688019a715eccaf81962c3e85868