Keep PublisherImpl alive until StreamPublisherImpls are alive
Summary:
We're going to keep track of the bytes buffered for a subscription or fetch in the `PublisherImpl`, so we need to ensure that it remains alive until all bytes have been delivered. I'm making the `StreamPublisherImpl` hold a shared_ptr to the parent `PublisherImpl` to ensure this.

This diff is just concerned with lifetime extension of the `PublisherImpl`. Buffering will come in a later diff.

Reviewed By: afrind

Differential Revision: D70989712

fbshipit-source-id: e0fe808bf14bfb74e5d2e544372f2f933bcdb4b9