Keep StreamPublisherImpl alive until all bytes delivered or cancelled
Summary:
We're going to keep track of the bytes buffered for a subgroup in the `StreamPublisherImpl`, so we need to ensure that it remains alive until all bytes have been delivered. I'm making the `StreamPublisherImpl` hold a shared_ptr to itself, so that it keeps itself alive while callbacks are still remaining.

This diff is just concerned with lifetime extension of the `StreamPublisherImpl`. Buffering will come in a later diff.

Reviewed By: afrind

Differential Revision: D70989452

fbshipit-source-id: 0807a6ef8af9d36a38f153e57f13351fb91ac70e