Fix nullptr dereference in byte buffering logic
Summary:
The following sequence of events was problematic:
1. User calls `StreamPublisherImpl::endOfTrackAndGroup` in order to end a stream.
2. We'd send an object header with `END_OF_TRACK_AND_GROUP`, and set the delivery callback to be `this` object (i.e. the `StreamPublisherImpl`).
3. Further downstream of `StreamPublisherImpl::endOfTrackAndGroup`, we'd call `StreamPublisherImpl::onStreamComplete`, which sets `publisher_` to `nullptr`.
4. Later on, when the data in that object header had been acknowledged, we'd attempt to call `publisher_->onBytesUnbuffered`, but `publisher_` would be `nullptr`.

To fix this, I modified `StreamPublisherImpl::onStreamComplete` so that it doesn't nullify `publisher_`, and instead introduced the variable `streamComplete_` in order to track stream completion.

However, this resulted in a heap-use-after-free in `TrackPublisherImpl::onStreamComplete` when we tried to erase from the `subgroups_`. This could result in the destruction of the only reference of the last `SubgroupPublisherImpl` that has a shared_ptr reference to this `TrackPublisherImpl`, causing the destruction of the map while erasing from it. I used a `shared_from_this` keepalive in order to prevent this from happening.

Reviewed By: afrind

Differential Revision: D71987182

fbshipit-source-id: a8f1b0eacc80febb811964fde193c0667f38ead5