fix cmake xxhash discovery
Summary:
Fix fb303 OSS builds on macOS that failing to find xxhash.h.

Problem was that thrift cpp lib now needs xxhash but it wasn't declared in the cmake rules, which shows up as a problem when system xxhash is not installed.

CustomCompressorRegistry.cpp was also missing from cmake build resulting in link errors once the xxhash.h discovery was fixed.

Reviewed By: markbt

Differential Revision: D72452460

fbshipit-source-id: 6bdb8e0e3961529f04f9d9d29ba0daeff2e7ff2a