WIP: MoQCache (#46)
Summary:
This is a simple in-memory cache for MoQ Tracks and Objects.  It currently doesn't have any maximum size or eviction behavior, so shouldn't be enabled in long-running relays.

The basic caching data structures are how you would expect - a map of tracks, each track has a map of groups, each group has a map of objects.  Objects that don't exist (eg gaps) are cached, but omitted in FETCH responses.

When a subscribe is active, a SubscribeWriteback filter is attached, that populates items in the cache.  When a fetch is active, a FetchWriteback is attached.  The FetchWritebacks are also managed in an interval set preventing more than one fetch for the same object upstream.

Fetch requests search the cache and go upstream for objects with unknown status.

Pull Request resolved: https://github.com/facebookexperimental/moxygen/pull/46

Differential Revision: D71044480