Support AUTHORIZATION_TOKEN
Summary:
I also deleted the old AUTHORIZATION param.

This leverages the MoQTokenCache class, one in the parser, for received tokens and one in session, for sending tokens.  For each API with parameters, we "aliasify" the parameter list:

If the token is marked 'DontRegister', send as literal
If the token is larger than the MAX_AUTH_TOKEN_CACHE_SIZE, send as literal
If the token is in the cache, USE_ALIAS
Othweise evict from the token cache using LRU, adding DELETE params until there's enough space; then REGISTER

Note it's possible to insert and evict an item from the token cache in the same parameter set.  It's important to copy the values out.

Reviewed By: sharmafb

Differential Revision: D73871904

fbshipit-source-id: 7f2260c6445ecd98c0b731f78f4782daeb47a5e2