Delta encode Object IDs within Subgroups
Summary:
The Object ID Delta + 1 is added to the previous Object ID in the Subgroup
stream if there was one.  The Object ID is the Object ID Delta if it's the
first Object in the Subgroup stream. For example, a Subgroup of sequential
Object IDs starting at 0 will have 0 for all Object ID Delta values.

Reviewed By: afrind

Differential Revision: D82571092

fbshipit-source-id: 6aba15422c23a44e1019fb4d82e6e55620cf5534