Add a MoQServer::start API
Summary: We need to have a start API because there's some initialization we'll have to do before we actually start running the MoQServer. For example, setting the Fizz Context, stats, and takeover-related information (such as the host id and process id). I'm going to add setters for this, and enforce that start is called after that.

Differential Revision: D82172899

fbshipit-source-id: 5394dd8c429534f7188306278ae885628e1b78f1