Search upstream subscribes for fetch on relay
Summary:
If a server only publishes a track (doesn't announce), it wont be present in the relay's announce tree. In this case, we just need to check the upstream subscriptions.

Note: There's an error with the dateserver currently where it doesn't update the groupID to >0 until it crosses the first group boundary, so if you launch the date server and then fetch from it shortly after it will likely fail.

Reviewed By: sandarsh

Differential Revision: D83174628

fbshipit-source-id: ebd7017e32aa38af7e0690e5bfd33389c8b0b7e6