Refactor FetchIntervalSet to use group-based intervals
Summary:
This refactor is a part of the effort to support descending order in subscribe/fetch.

Fetch interval set is used to track the fetches that are currently active. Previously, because there was no group based separation, fetch request that span multiple groups create a single entry. When a particular object was successfully fetched, the interval start would just be incremented. This approach will not be able to support descending order in fetch, as going in ascending fashion is tightly coupled. As part of this redesign and decoupling, the fetch interval will be broken into respective groups. That will enable us to iterate both in ascending and descending order per group (while the objects will always be processed in asc order, as part of spec).

Added more tests to include scenarios that span across groups.

Reviewed By: sharmafb, afrind

Differential Revision: D80851373

fbshipit-source-id: b27d3164af4a4be20d9003dab64ab2825161fc5f