Updates to trackStatus
Summary:
Turns out TrackStatus is used in a lot of different places so this is kind of a huge diff.

Stops sending track status requests after goaway (https://github.com/moq-wg/moq-transport/pull/1105/files)
Rewrite track status to mimic subscribe interface (https://github.com/moq-wg/moq-transport/pull/1015/files)

For TRACK_STATUS, I simply alias TrackStatus = SubscribeRequest, since every legacy field of TRACK_STATUS_REQUEST is included in Subscribe. However, the new TRACK_STATUS_OK message (previously "TRACK_STATUS") has legacy fields not included in SubscribeOk, so for that I copied it to its own struct, but in the future we should be able to alias as well.

There's still some work to do on unifying the TrackStatus and Subscribe logic in a few places (particularly in the date server sample code, which I just put a placeholder that is type compatible), but this should address most of the needs for interop.

Reviewed By: sharmafb

Differential Revision: D82342984

fbshipit-source-id: d8981171810189326f74ab97e105a130a886b74c