Add support for a frame by frame flv stream processing in flv parser
Summary:
Moxygen flv streaming client and receiver client deals with files. But, we have other places where we receive FLV stream frame by frame and it needs the same exact sequential stream processing to assign frameId, determine group boundaries, etc. Moved FlvSequentialReader functionality into a FlvStreamParser that deals with frames and modify the reader to provide a frame from the file rather than a file reader directly. This is largely just a refactor of the existing implementation with the additional minor logic to deal with frames that come with padding bytes at the end.

With this change:
`FlvCommon` provides the functionality to parse `ONE` flv tag (video, audio or script) from a given IOBuf.
`FlvStreamParser` provides the functionality to parse a stream of flv frames in the form of IOBuf. It tracks `frameId` or `seqId` of the video and audio frames in the stream. Previously, `FlvSequentialReader` was doing this assuming it is parsing a flv file.
`FlvReader.cpp` now opens the file, utilizes `FlvStreamParser` to generate a series of Flv tags.

Reviewed By: afrind

Differential Revision: D82170710

fbshipit-source-id: b13d589b7b9c123ac481836e66aef6efb95d33e5