Reduce subscribeDone confusion and fix UAF
Summary:
The publisher could get a UAF when trying to publish after session closure.

To the dismay of Mrs. McCave:

1. Removed PublisherImpl::subscribeDone -- this was a think wrapper around the MoQSession::subscribeDone
2. Renamed MoQSession::subscribeDone to sendSubscribeDone, since that's what it's doing
3. Renamed TrackReceiveState::subscribeDone to processSubscribeDone, since that's what it's doing

The only method still called subscribeDone is the one in TrackConsumer that must be overridden.

Fix the UAF by clearing session_ in terminatePublish.  Added a check in subscribeDone as well (any other publish function would call getWebTransport(), which checked it).

Reviewed By: sandarsh

Differential Revision: D86000424

fbshipit-source-id: 709e4d6cf3e1c33f9b65799f310e1f99c6d78c50