Remove MoQRelay::removeSession spaghetti
Summary:
This method was doing linear searches of state to ensure that a session closed by the peer was purged from all the data structures.  However, MoQSession now correctly invokes all state terminating callbacks that remove the state in O(1) or O(depth) time.

One of the key helpers was adding a TrackConsumerFilter that intercepts the receiving done callback.  This resets the upstream subscription handle and makes onEmpty simpler as well.

Reviewed By: sandarsh, sharmafb

Differential Revision: D86773281

fbshipit-source-id: 449a9c86f651bea90c8eb1ad3a4a65ba4508f127