Enable PGO + LTO in Sandcastle job
Summary: To make these work we need to make LLVM's runtime profiling library/llvm-ar or a newer version of GCC available on Sandcastle hosts. I've gone with making GCC-14 available as this also matches the cibuildwheel build we have on GitHub and so will help keep our internal and external builds consistent.

Reviewed By: alexmalyshev

Differential Revision: D87987135

fbshipit-source-id: 5670450efb6fdb6520c828f3ee7ea504a64b79dc