Short circuit read loop if callback returns error
Summary:
This adds a ParseResult return value to a couple more codec callbacks, and checking in MoQCodec to return early on ERROR_TERMINATE.  BLOCKED is not a valid error from the callbacks other than onSubgroup.

So if readStreamData returns 100 objects worth of bytes, and the first callback returns TERMINATE, we don't continue parsing the rest.  We eliminated the error_ member of the ObjectStreamCallback in MoQSession, and just check the onIngress return value.

Reviewed By: sharmafb

Differential Revision: D87907663

fbshipit-source-id: f2c367e1c966cd7af73cff73ccad0ba1b757f466